DBValue Wrapper
This commit introduces the DBValue wrapper, a thin utility around a type-erased object. This abstraction provides a
domain guarantee that the wrapped object is suppoted by the framework and can actually be stored in a database,
centralizing the type-checking code at a single source. This struct should be used in the various Kvasir APIs for safety
and expressiveness.

Additionally, this commit adds a couple of testing helpers. One is a pair of attributes that enable/disable test methods
based on the DEBUG/RELEASE status of a build. The second is a family of helper functions for evaluating the full
equality API of a class (i.e. Object.Equals, IEquatable<>.Equals, operator==, operator!=, and GetHashCode) on a pair of
inputs. The former will be useful when testing precondition checks, while the latter is useful for efficiently writing
tests that exercise user-defined types' equality interfaces.