DBType 'Smart' Enumeration
This commit implements DBType, an instance-limited struct that represents a semantic data type. This struct represents
every semantic that the framework will support: if a .NET Type has no DBType equivalent, then it is fundamentally
unsupported. Like an enum, the instanes of DBType are accessed by a pseudo-namespaced identifier; unlike an enum, there
are available member functions and metadata can be attached (though none currently is).

DBType does not provide any immediate connections to provider-specific enumerations indicating actual storage, nor does
it support specification of auxiliary informaiton such as precision or valid ranges. This is intentional: this type is
intended to describe the semantic of the Schema rather than anything about implementation in the back-end database.