ConceptString: A Strongly-Typed Domain String
This commit adds ConceptString, a light wrapper around a raw .NET string intended for representing application and/or
library domain concepts. The utility provides the barebones functionality of equality comparisons, ordering comparisons,
simple type conversions, and iteration; additional functionality is left to the derived class based on the specific
domain concept at hand. This utility will be used to represent Schema component names (among other things) within
Kvasir.

Along with the ConceptString, this commit also modifies the implementation of the FullCheck equality helper and adds
FullCheck comparison helpers in a parallel vein. The implementation changes primarily address shortcomings in the use of
dynamic typing, especially in the case of null.