IField Interface, FieldName, and IsNullable Enum
This commit adds the IField interface, which will be used for all concrete Field implementations in Kvasir. This
interface is minimal, containing only data that is relevant to a Field's schema declaration. This interface is likely to
expand slightly in the future, particularly as it pertains to generating SQL.

Auxiliary pieces introduced in this commit are FieldName (a ConceptString denoting the name of an IField) and IsNullable
(a strongly typed Boolean yes/no).