Enum IsValid Extension Method
Adds an extension method on types constraint to Enum that checks it an enumerator is valid. This is inteded to be used
as a defensive check to guard against arbitrary numeric casts being passed to APIs that expect defined constants. The
utility is currently employed in BasicField and EnumField for the Nullability input enum on construction.