Initial SQL Transcription Interfaces
This commit introduces the initial interfaces for the Transcription Layer, which is responsible for converting from the
Schema representation of various components (e.g. Fields, Tables, Constraints, etc.) into syntactically valid SQL.
Specifically, this commit focuses on declaratory SQL that is used to build up CREATE TABLE statements. We use a modified
factory pattern and dependency injection to group provider-specific components into a collection of interfaces that can
then be handed off to a Schema component; the component then obtains the appropriate SQL generator from the collection
to create a snippet of valid SQL.

Note that this commit does ont implement any specific generators; rather, it provides the interfaces necessary for the
various Schema components to leverage via dependency injection. The generator collection interface will grow as
additional components are added that need their own dedicated generator types. The actual implementations of these
interfaces will be left as a future exercise.