Check Constraint Clauses
This commit provides the fundamental building blocks for all CHECK constraints that will be supported by Kvasir. These
building blocks are offered as a Clause hierarchy, supporting four atomic units: nullity checks, inclusion checks,
constant value checks, and cross-Field value checks. From these units, arbitrarily complex Boolean expressions can be
built, ranging from relatively simple AND and OR to more complex XOR, IF-THEN, and IFF. The SQL generation interfaces
have been extended to interact with the classes in the Clause hierarchy.

Although this commit implicitly enables support for the arbitrarily complex CHECK constraints, only a simplified subset
will be initially supported by the framework. This will be controlled via the APIs and attributes exposed to the user as
an input mechanism; the purpose of building out the entire hierarchy is to allow for seamless extension in the future.