Rebrand Declaration Factories as Declaration Builders
The interfaces responsible for generating SQL declarations that are aggregated in a provider-specific collection and
thereby exposed to Schema components conform more to the Builder pattern than to the Factory pattern, so they've been
renamed accordingly. The IConstraintDeclBuilder did not undergo any significant changes; the IFieldDeclBuilder was
refactored into separate setters for the various pieces as is befitting a proper builder. The previous extension method
was deleted as well. All builders now fall under the umbrella of the IDeclBuilder interface, which exposes Build() and
Reset() methods only.