Check Constraint Schema Component
This commit adds the Schema component representation of a CHECK constraint, which is a thin wrapper around a Constraint
Clause. A CheckConstraint has an optional name and has the same declaration generation API as the classes in the Field
hierarchy, delegating to the constituent Clause to do the actual work. The CheckConstraint class will be used to
represent all varieties of CHECK constraint, as the complexity is buried in the syntactic representation of the Clause
class.