Primary and Candidate Keys
This commit builds the Schema representations of Candidate Keys (for UNIQUE constraints) and Primary Keys. The latter is
simply a called-out instance of the former, with the only non-semantic difference (other than different C# types) being
in the generated constraint declaration. The classes are fairly minimal and will be used by the Table implementation,
which will require exactly one Primary Key and zero or more Candidate Keys.