Add ToString() Override for DBType
This commit adds an override of the object.ToString() method for DBType. To facilitate this without adding extra members
to the DBType class, a static array of string forms is used with the existing byte ID serving as an indexer. Due to
static initialization ordering rules (which I don't fully understand at this time), this requires that the previous
in-line initializations for the DBType pseudo-enumerators (e.g. public static ... { get; } = ...;) be moved to the
static constructor.

This commit resolves GitHub Issue #3.