Major Refactor: Add ForeignKey & Table, Upgrade to C#9, and Extract Out Cybele/Atropos
This commit executes a MAJOR refactor of the entire codebase, reflecting a philosophica shift in the way the repository
is maintained. The repo is now explicitly separated between three projects: Kvasir (for the ORM), Cybele (for core
utilities), and Atropos (for unit testing utilities). Eventually, the latter two will be extracted out into their own
packages; each has its own top-level namespace, is documented separately, and will have its own README.

Aside from the structural changess, many implementation changes were made. The entire codebase was updated to C#9, which
requires the net5.0 framework. The following modifications wre made to the actual implementation:
  * Schema Layer components are now records, with debug-only error checking
  * Guard Clauses are now used for error checking where applicable
  * A new ComponentName class has been added as a central location for name validation
  * Properties of type IEnumerable<T> have been minimized, preferring IReadOnlyList<T> or something similar
  * Within the Clause subsystem, negation is handled differently and a dedicated XOR class now exists
  * ConceptString now supports substringing via range-based indexing
  * The type-value validity check has moved from DBType to DBValue
  * Several new utility functions in Cybele, particularly targeting IEnumerable<T>
  * Most classes and interfaces are now public rather than internal, with constructors and some methods the exception
Additionally, this commit adds the full implementation for ForeignKey and Table, the latter implementing the ITable
interface. The former supports behaviors for ON DELETE and ON UPDATE.

Note that this commit does not include unit tests for the refactor code: that will be forthcoming in a subsequent
commit.