Always Build Unit Test Project in Debug Mode
The unit test suite for the Kvasir solution leverages FluentAssertions, an open-source NuGet package for performing test
expectations in a naturalistic way. This package uses stack-backtracking to locate the 'variable under test' when
producing error messages; as such, they recommend only building projects using the package in Debug mode. This does not
affect the projects under test, which are still built in Release mode for the Release configuration; only the unit test
project is affected.

This commit also hard-codes the Coverlet package into the UnitTests project and removes the manual add from the Travis
.yml file.