Documentation Generation via Doxygen, Sphinx, and Breathe
So this is basically working now thanks to the CSharp domain provied by roger-barton on GitHub. There are some things that need
to be improved:
  > The current single Kvasir RST is way overloaded. Some of that stuff needs to be broken up into separate pages (e.g. one per
      subproject, one per namepace, etc.)
  > Warnings/Errors should probably be disabled when executed from the command line, because there are a LOT of false positives
  > Will have to figure out how to integrate this into the build pipeline, potentially; would require the necessary packages

For now, going to keep this on a side branch as it's not the most important thing. I can revisit this later, once all of the
core functionality is complete and I have a working prototype.

UPDATE [7/10/2022] - Getting a bunch of errors now when I run 'make html', complaints are about invalid function signatures.
Not sure what's going on.