Extraction Plan
This commit adds the logic for building ExtractionPlans, which are pipelines for pulling values out of an input object.
The IExtractionPlan instances leverage DataConverters to perform any necessary transformations and return collections of
ready-to-write DBValues. Note that the ExtractionPlans are not responsible for understanding how to create a plan: that
will be the responsibility of the translation layer, with these components solely responsible for the execution pieces.

Additionally, this commit adds two new extension methods (IEnumerable.AllSame and Type.IsInstanceOf) that were needed by
the ExtractionPlan implementations. Lastly, a bug in the DataConverter logic related to the conversion of null when a
value type is expected was fixed, and the unit tests were accordingly updated.