Reconstitution Plan
This commit implements the Reconstitution Layer, which is a complement to the Extraction Layer implemented in the
previous commit. This layer is responsible for modeling the process by which a row of data from the database is turned
back into a CLR object through two phases: creation (e.g. calling a constructor) and mutation (e.g. setting property
values). As with the Extraction Layer, the current tools available in the Reconstitution Layer are intentionally limited
but specifically designed to be forward compatible with other mechanics, such as factory functions and arbitrary method
invocations.