StickyList
A StickyList is a list-like data structure in which each element has a 'stickiness' that defines whether or not its
position in the list is fixed. A non-sticky element can occupy any available slot, whereas a sticky element must ocupy
the slot at which it was inserted/added. This can lead to gaps in the list, which are filled by non-sticky elements
whenever possible (such as on addition, or when a gap arises due to a removal).

The StickyList will be used for ordering columns during Translation, as some Fields will be pinned to specificy column
indices while others will be flexible. That being said, the StickyList itself is intended to be general-purpose: for
example, it throws exceptions on invalid or noncomformant input and exposes certain query APIs that may not be useful
for the specific Translation use case.