Move DataConverter Class to Cybele
This commit moves the DataConverter utility class out of the Kvasir project and into the Cybele project. In the course
of designing the annotation system for Kvasir, I realized that the current DataConverter utility is insufficient for
user-facing needs: the annotation input will have to be a standalone type, which will necessitate clients writing their
own converter class anyway. This forced a re-evaluation of the existing helper, and I reached the conclusion that it
will be a useful internal representation of the conversion mechanic much the way it currently is being used for
Extraction and Reconstruction. Because of this, and because the current implementation has no dependencies on anything
else in Kvasir, it made sense to move it to Cybele as a pure utility.