Field- and Table-Level Annotation Attributes
This commit introduces a grand total of 28 attributes that can be applied to an entity class or a property thereof to
annotate the corresponding database construct. These attributes range from overriding the default naming behavior to
instituting constraints on values. Most constraints can be applied to properties of aggregate-type Fields using the
'nested' variety, in which a dot-separated path is provided on construction.

This is not necessarily an exhaustive set of all annotations that will be available through Kvasir. Rather, it is the
baseline set for the beta release. Enough flexibility has been built into the system to allow for creation of
arbitrarily complex CHECK constraints, which should suffice in the interim.