RelationList<T>
This commit implements RelationList<T>, the first collection type of the Kvasir framework. RelationLists are unordered
collections that track mutating operations (insert, append, overwrite, remove) and can expose the current 'state' of
their elements for appropriate interaction with a back-end relational database. The API of a RelationList is identical
to that of a standard List<T>, with the addition of a few explicit interface implementations of the new IRelation API
(which is internal, and therefore hidden from users). RelationLists should only be used with immutable elements, and
while they technically permit duplicates, this should be strenuously avoided.

This is the first of at least three planned Kvasir collections: the enxt two are a set-like collection and a map-like
collection. They will follow the same patterns (implement the IRelation API plus the APIs of the standard collection
they mimic), which will collectively allow for seamless integration with the Extraction, Reconstitution, and Translation
layers of the framework.