RelationSet<T>
This commit implements the RelationSet<T>, the second collection type of the Kvasir framework. RelationSets are
unordered collections that track utation operations and disallow duplicate values; they can expose the current 'state'
of their elements for appropriate interaction with a back-end relational database. The API of a RelationSet is identical
to that of a standard HashSet<T>, with the addition of a few explicit interface implementations of the new IRelation API
(which is internal, and therefore hidden from users). RelationSets should only be used with immutable elements.