RelationMap<TKey, TValue>
This commit implements the RelationMap<TKey, TValue>, the third collection type of the Kvasir framework. RelationMaps
are collections of key-value pairs that track mutation operations and disallow duplicate values; they can expose the
current 'state' of their elements for appropriate interaction with a back-end relational database. The API of a
RelationMap is identical to that of a standard Dictionary<TKey, TValue>, with the addition of a few explicit interface
implementations of the new IRelation API (which is internal, and therefore hidden from users). RelationMaps should only
be used with immutable value types.