RelationExtractionPlan
This commit adds the RelationExtractionPlan, a type dedicated to pulling an IRelation off of an Entity object and then
iterating over its contents to extract data rows. This plan, like the EntityExtractionPlan that it relies on to operate
on relation elements, does not implement any specific interfaces: it is intended to be used directly, and relies on
dependency injection design principles for flexibility. Executing a RelationExtractionPlan produces a dedicated record
type that amounts to a data triplet (DELETE, UDPATE, INSERT); this data will then be processed by the yet-unwritten
component responsible for interacting directly with the database.