Add Repopulation Interface to IRelation
This commit adds a function to the IRelation interface for adding an item into an IRelation during the reconstitution
process. This API is called 'Repopulate' and takes a type-erased object, as the IRelation interface is not strongly
typed. However, the actual item itself will have been strongly reconstituted, allowing for a direct cast. Repopulated
items are added in the Saved state, mitigating the need for any canonicalization.