RelationReconstitutionPlan
This commit adds the RelationReconstitutionPlan, the conceptual opposite of the RelationExtractionPlan. This construct
takes a collection of rows and, for each one, reconstitutes the row as some kind of sub-Entity from which the Relation
is then repopulated. Like other pieces in the Extraction/Repopulation space, the plan is heavily interface-based: a new
IRepopulator interface defines how to repopulate a Relation, with a single property-read implementation available at
this time.