Clean Up Implementation of IsInstanceOf
This commit cleans up the implementation of the IsInstanceOf extension method in Cybele, as well as improving the
documentation and adding some additional unit tests for robustness. The primary changes to the method are fixing an
issue whereby open generics were improperly considered in certain situations, and the elimination of an unnecessary
recursion. Most of the new tests highlight scenarios in which the method should return false that may be unexpected (but
are now thoroughly documented).

This commit resolves GitHub Issue #26.