Fix In-Source Documentation re: Extrinsic vs. Intrinsic Transformations
An _Intrinsic Transform_ is one that is applied natively and automatically by the Framework. The most common form of
such a transform is to convert from a C# enumerator into a string. Other examples would be converting between integer
types for back-end RDBMSes that don't support all integer sizes. Whether or not an Intrinsic Transform is performed
depends solely on the value's data type and the specifics of the back-end storage, nothing else.

An _Extrinsic Transform_ is one that is applied based on the user's explicit designation. Converting a C# enumerator
into an integer is an example of an Extrinsic Transform, even if the implementation is entirely provided by the
Framework. The DataConverter attribtue is the most common avenue of enacting an Extrinsic Transform.

It is possible to have both an Intrinsic and an Extrinsic Transform applied to a single value. For example: the value of
an enumeration-type Field may be Extrinsically transformed into its numeric backing type, which is then Intrinsically
transformed into the corresponding unsigned form. When extracting data, the Extrinsic Transform is applied first; when
reconstituting data, the Intrinsic Transform is first.

This commit resolves GitHub Issue #27.