Rename EntityExtraction and EntityReconstitution Plans
This commit performs a pair of renames: EntityExtractionPlan -> DataExtractionPlan and EntityReconstitutionPlan ->
DataReconstitutionPlan. This is to reflect the fact that these plans do not fully extract or reconstitute an Entity but
rather a single row's worth of data from an object. We will be creating an actual EntityExtractionPlan and
EntityReconstitutionPlan shortly to get operate on all of an Entity's data.