Fix Build Status Badges on README
This commit fixes the build status badges in the README file, which were not updated in the previous commit when we
mirgarted to GitHub Actions. This necessitated splitting the workflow file into two, one for Ubuntu builds and one for
Windows build, as the badge URLs are based on the action name.