Improve Unit Test Coverage by Adding New Tests
This commit adds a bunch of new tests, and extends a handful of existing tests, to improve the coverage of the combined
projects. The majority of the coverage gains came from adding tests for missing functions on the various Relations,
primarily explicit interface instantiations. The serialization constructor for the KvasirException is also now tested,
as is the error condition for the Kvasir IDataConverter. A few Debug assertions were also split from being compound
expressions into multiple expressions to reduce the number of uncovered branches.