Rename Row Alias --> DBData
This commit changes the Row alias to DBData, which is more descriptive and less prone to confusion. The old alias was
used in places where the ReadOnlyList may have been a full database row, or a partial database row, or either a full or
partial database row. This made the alias functionally useless.