New ObjectCreator to Match Existing Entities by Key
This commit adds a new IObjectCreator implementation, the ByKeyLookupCreator, that treates the raw values provided to
the Execute function as a key of an existing Entity. That key is then compared against a collection of possible matches,
which is created by invoking a 'factory' function. A match is required, as the absence of one would be tantamount to a
violation of a foreign key constraint in the back-end RDBMS. The use of a factory allows for easy testing, external
control of the possibilities, and eliminates the need for globals or statics.