Allow ByConstructorCreator to Identify All-Null Values
This commit changes behavior of the ByConstructorCreator in the event that all of the values provided are DBValue.NULL.
Because DBValue.NULL is always a sentinel value in Kvasir, there are two situations in which this could happen: with an
optional aggregate Field that is omitted, or with a required optional Field that contains all optional Fields that are
omitted. In the former case, we want to produce the C# value 'null' instead of invoking the constructor; in the latter
case, we want to invoke the constructor. To facilitate this, the ByConstructorCreator's own constructor now takes a
Boolean operand that controls the behavior when it encounters all DBValue.NULL.