Make Transcription More Generic & Flexible
This commit does a small facelift on the interfaces exposed by the Transcription subsystem. Specifically, those
interface now operate on a generic parameter that defines the declaration type being used rather than automatically
using SqlSnippet. The rationale for this is simple: Kvasir should support any back-end storage mechanism, including
those that don't operate via SQL (e.g. a CSV file store). This change doesn't affect any of the actual logic, as the
individual Schema components were simply using the builder APIs and then calling the Build() method, returning the
result; this behavior is fully unchanged.

Note that because of the annoying lack of type aliases in C#, the ITableDeclBuilder has to be generic over all of the
individual declaration types, since it refers to them in its API. For obvious reasons, the IBuilderFactory also has this
property.

The unit tests are unchanged: where necessary, SqlSnippet was used as the generic type parameter to limit mimic the old
behaviors.