Remove Data Conversion from Extraction/Reconstitution Steps
This commit removes the responsibility of performign ANY data conversion/reversion from the PrimitiveExtractionStep and
the PrimitiveObjectCreator. The transforms are still performed, but at the higher level; this is because, fundamentally,
the steps in the plans for Extraction and Reconstitution are predicated on the target type, and should be fully divorced
from the source Entity type. This change, for example, makes the DecomposingExtractionStep something that can be reused
across different Fields, even if those Fields wish to apply different extrinsic transforms. The Translation layer will
have to compound the intrinsic and extrinsic transformation logic into one that is applied either at the end (for
Extraction) or the beginning (for Reconstitution). This also may have a slight performance benefit, eliding the need to
do multiple identity conversions when no transforms are requested.