PropertyChain
This commit adds the PropertyChain utility class to Cybele. This is a data structure of sorts that represents a sequence
of readable properties that supports read operations on source objects. When a value is extracted, each link in the
chain is executed via reflection and the resulting value is used as the source object for the subsequent link. (Any null
values immediately terminate the descent.) The PropertyChain is intended to support certain data extraction needs in
Kvasir, particularly for dealing with Relations stored on Aggregate-type Fields and for reconstituting properties on
read-only Aggregates. Though this is the intended use case, the PropertyChain class is still more generally versatile:
it supports both public and non-public properties, and both instance and static properties.