Allow ReadPropertyExtractor to Read from PropertyChain
This commit updates the constructor of the ReadPropertyExtractor to support reading from a PropertyChain instead of just
a single property. This primary need for this is specifically obtaining Relations that are nested in Aggregate-type
Fields, as the RelationExtractionPlan accepts an IFieldExtractor which previously could only read one property deep. The
plan is still to use the DecomposingExtractionStep to get the data out of Aggregate-type Fields, because this will allow
for reusability.