Extend Reconstitution Components to Work with PropertyChains
This commit adds an IFieldExtractor argument to the SetPropertyMutationStep and the FromPropertyRepopulator to enable
them to function properly if the target property is not actually on the entity source object. The new constructor
argument controls how the target object is obtained from the initial source, which allows these classes to interact with
content nested in Aggregate-type Fields. This is necessary to support Relations modeled in this fashion, as well as
read-only Aggregate-type Fields that expose writeability on their constituent properties.