New Attributes: ExcludeNamespaceFromName and PathSeparator
This commit adds two new attributes that can be used to control how names for components are determined by Kvasir. The
first is the ExcludeNamespaceFromNameAttribute, which directs Kvasir to ignore the C# namespace when creating the name
of a POCO entity's primary table. The second is the PathSeparatorAttribute, which directs Kvasir to use a specific
character to separate portions of a property gleaned from an Aggregate.