Scalar Property Translation
This commit is the first major milestone in the implementation of the Translation Layer of Kvasir. With this commit, the
Framework is able to properly translate any Entity Type consisting solely of scalar properties into a data model. Nearly
all annotations are supported; the only one that isn't is the CheckAttribute. Primary Key deduction is fully
implemented, as is Candidate Key identification. All naming rules are recognized for both Fields and for Tables. All
pertinent error checking is present, with error messages contextualized to explain where the error is coming from. More
than 140 unit tests ensure that the logic is sound. The only changes to the specification were clarifications made; no
substantive content was modified.