Clean Up Heritability and Error Management of Annotations
This commit makes two broad sets of change to the Annotations exposed to users by Kvasir. The first is to make all
custom attributes non-inherited. Kvasir currently doesn't support inheritance with Entity types (in the sense that
inherited Fields, from interfaces and/or base classes, are not included in the data model). So, having the attributes be
inherited doesn't really make sense.

The second change is larger: outside of things like null checks, attribute constructors should NEVER throw an exception.
The reason for this is that the Translation Layer will be able to contextualize the error messages where the annotations
themselves cannot: we want to be able to tell users what Entity type and what property triggered the invalid annotation
error (even though the debugger probably will do this anyway). To facilitate this, a few error checking operations were
just outright removed, and others were elided by changing the exposed data type of the information carried by the
attribute. For the CheckAttribute (and the CheckComplexAttribute) and the DataConverterAttribute, where the constructor
logic involves creating an object of an expected interface via reflection, erroneous user input is detected and an error
message is stored on the attribute that the Translation Layer can later access systematically.

In the course of the latter change, a small but meaningful number of logical edge cases in the error handling were
identified as being untested. These didn't show up in the code coverage report because they were triggered by a
single-line reflection invocation that could also trigger other errors that were actually tested. Tests for the new
scenarios have been added.