Fix Up some Annotation Property Visibilities
This commit changes the visibility of the Path properties on the 8 attributes that can be made to apply to nested Fields
from internal to public. This is because a quirk in C# means that even if the initter of a property is public, it cannot
be used during construction if the getter is non-public. Without this change, the Path properties would be useless.

This commit also makes the UserError properties for CheckAttribute and DataConverterAttribute read-only outside of the
class (i.e. a private setter).