Better Handling of Anonymous UniqueAttributes
This commit adds two new features to the UniqueAttribute annotation: a globally reserved prefix for names and a
per-instance property indicating if a name was provided on construction or not (i.e. if the annotation is 'anonymous').
These features are necessary so that Translation can identify if a user has specified an explicit name that clashes with
a name used for an anonymous Candidate Key, which would otherwise cause the single-Field key to have multiple Fields.
The specification has been updated accordingly.