Add Ability to Create Identity DataConverter from Runtime Type
This commit adds a new static API to the Cybele.Core.DataConverter class that enables the creation of an identity
converter from a runtime Type object. Previously, the only API to create identity converters required the explicit
specification of a generic type parameter, which cannot be done by Translation (which needs to leverage reflection
heavily). The new API is, in all other ways, identical to the existing API.