Clear Out Kvasir Settings Object
This commit removes the single trait, DateTimeStyle, from the Kvasir.Core.Settings object. The original expectation was
that this value could be used to control how strings are pared into DateTimes (e.g. YYYYMMDD vs. MMDDYYYY vs. DDMMYYYY
etc.), but this is not actually the case. As such, it doesn't really serve any purpose: the Translation Layer can't
actually use it for anything.

The Settings object itself still exists, as a form of forward compatibility where we can add customizations later if we
find the need.