IsInherited Extension for Methods
This commit adds a new extension method for MethodInfo and ConstructorInfo that determines if the method is inherited or
not. An inherited method is any one that was not first declared in the type through which it was accessed via
reflection; this includes non-virtual methods defined in base classes, overrides of virtual and abstract methods, and
all interface implementations (implicit and explicit). Static methods are analyzed in the same way. Constructors are
implicitly not inherited.