Revamp Way Constraint Annotations are Handled
This commit redoes the design of the constraint-imposing annoations in the Kvasir.Annotations namespace. Previously,
each type of constraint was afforded its own annotation that derived from a base ConstraintAttribute. The base class
maintained internal member variables for an optional function, an operator (inclusion or comparison), and an anchor. It
then exposed a MakeConstraint(...) API that took an IField and a DataConverter and created the corresponding Clause from
the aforementioned member variables. While this was an elegant solution that mimicked the constraint grammar, it had a
fatal flaw: it inhibited any kind of robust error checking during Translation, because the user data was inaccessible.

Given that, and with a better understanding of how Translation is going to be implemented, the base ConstraintAttribute
has been removed. It has been replaced with three new base classes specific for each subgroup of constraint annotations:
ComparisonAttribute, InclusionAttribute, and StringLengthAttribute. These base classes expose the operator and anchor
(appropriately as either a scalar or an IEnumerable) and perform no error checking on the input values. This means that
Translation does not have to have explicit carve-outs for each member of the subgroup, though it does have to make sure
to handle the subgroups correctly. The burden of all data validation (including parsing strings into DateTimes and
Guids) shifts to Translation, mirroring the way the [Default] annotation is handled.

Another consequence of this refactor is that the TransformUserValues utility is no longer necessary and, as such, has
been removed completely. It was actually slightly buggy anyway (it didn't apply the provided DataConverter to
non-enumeration scalars), and it _also_ wouldn't not have afforded any opportunity to do error checking (particularly
for DateTime/Guid parsing) during Translation. Its logic is still available in Git history, and it may be a useful
reference when implementing the Translation Layer for constraint-imposing annotations.

As a final note, all non-base-class attributes have been declared as sealed. Previously, some were and some weren't;
this is just a homomgenization.