Translation of Entity- and Scalar-Scope Constraint Annotations
This commit implements the full Translation of constraints applied to Entity Types and to Scalar Properties into the
corresponding data model. Each existing flavor of constraint is fully supported, including custom user-defined ones via
[Check] and [Check.Complex]. The full complement of error reporting is also available, though there's a good deal of
code duplication among the error conditions and error messages that will likely need to be refactored. The existing
descriptors have been slightly extended to account for constraint translation, though no existing Translation logic
needed to be modified.

In the course of implementation this slice of the Framework, it became evident that the current back-and-forth design
whereby the top-level function requests translated information one piece at a time will not work in the long term. When
it comes time to deal with Aggregates and other non-Scalar property types, the current Translation logic will fall
slightly short, particularly because of the hard-coded impermissibility of multiple annotations and the requirement of
an empty Path. While the foundational algorithms are correct (as evidence by more than 200 unit tests), the general
control flow will need to be tweaked before the full set of envisioned features can be properly supported.