Remove Path Property from [DataConverter] Annotation
The [DataConverter] annotation no longer supports specification of a Path, as it causes complications when mixed with
constraints that operate on the post-conversion type of a property. If a Path'ed [DataConverter] annotation were applied
to an Aggregate Property and resolves to a nested scalar that is set to have a String Length constraint or a Signedness
constraint, what happens: the [DataConverter] changes the Field's type to something that no longer supports the
constraint. We could simply drop the constraint, but the same issue holds for arbitrary user-defined CHECK constraints,
which may presuppose a particular data type. To rectify this, we are just making it impossible by removing the Path from
the [DataConverter] annotation altogether.